move functions to migrate.rs
